switch to new spectrum convention
