+++ maria/tests/test_simulation.py
@@ -14,7 +14,7 @@ all_test_plan_configs = list(read_yaml(f"{here}/configs/test_plans.yml").values(
 
 all_instruments.pop(all_instruments.index("alma/ALMA"))
 
-n_sims = 8
+n_sims = 6
 test_instruments = np.random.choice(a=all_instruments, size=n_sims)
 test_sites = np.random.choice(a=all_sites, size=n_sims)
